Core: Move story-index helpers to shared instead of splitting csf guards
The csf-guards extraction kept breaking typecheck: its Story/Meta
signatures had to pick one module identity (source or bundled
declarations), and CI's check programs mix both, so some consumer always
saw two nominally different Story types (csf-factories.test.ts this
time).

The actual reason the preview-api store chunk sat in the Node CLI graph
was a layering wart: core-server imported sortStoriesV7 and
userOrAutoTitleFromSpecifier from inside preview-api's store, dragging
the store chunk — and via its barrel imports, storybook/test — into
every bootstrap. Those helpers are pure and isomorphic, so they move to
shared/story-index; core-server and common import them there, and
preview-api re-exports them from the new location, keeping its public
surface unchanged.

With that edge cut, the guards split is unnecessary: csf-factories.ts
returns to its next state (guards defined inline), and the preview-api
store files keep their ordinary barrel imports. Verified by CPU profile:
the tools CLI graph evaluates no storybook/test, @testing-library, or
csf barrel code.
