+++ code/core/src/preview-api/modules/store/csf/csf-factory-utils.ts
@@ -1,5 +1,4 @@
-import { isMeta, isStory } from '../../../../csf/csf-guards.ts';
-import type { Meta, Story } from 'storybook/internal/csf';
+import { type Meta, type Story, isMeta, isStory } from 'storybook/internal/csf';
 import type {
   Args,
   ComponentAnnotations,
