+++ code/core/src/preview-api/modules/store/csf/normalizeStory.ts
@@ -1,5 +1,5 @@
 import { deprecate, logger } from 'storybook/internal/client-logger';
-import { storyNameFromExport, toId } from 'storybook/internal/csf/csf-utils';
+import { storyNameFromExport, toId } from 'storybook/internal/csf';
 import type {
   ArgTypes,
   ArgsStoryFn,
