+++ code/core/src/preview-api/modules/store/csf/portable-stories.ts
@@ -1,5 +1,4 @@
-import { isExportStory } from 'storybook/internal/csf/csf-utils';
-import type { CleanupCallback } from 'storybook/internal/csf';
+import { type CleanupCallback, isExportStory } from 'storybook/internal/csf';
 import { MountMustBeDestructuredError } from 'storybook/internal/preview-errors';
 import type {
   Args,
