+++ code/core/src/preview-api/modules/store/csf/prepareStory.ts
@@ -1,6 +1,4 @@
-import { includeConditionalArg } from '../../../../csf/includeConditionalArg.ts';
-import { combineTags } from 'storybook/internal/csf/csf-utils';
-import type { CleanupCallback } from 'storybook/internal/csf';
+import { type CleanupCallback, combineTags, includeConditionalArg } from 'storybook/internal/csf';
 import { NoRenderFunctionError } from 'storybook/internal/preview-errors';
 import type {
   Args,
