+++ code/core/src/preview-api/modules/store/index.ts
@@ -8,6 +8,6 @@ export * from './csf/index.ts';
 export * from './hooks.ts';
 export * from './decorators.ts';
 export * from './args.ts';
-export * from './autoTitle.ts';
-export * from './sortStories.ts';
+export * from '../../../shared/story-index/autoTitle.ts';
+export * from '../../../shared/story-index/sortStories.ts';
 export * from './reporter-api.ts';
