+++ code/core/src/shared/story-index/autoTitle.ts
